CouponAPIを使う処理の更新だけやった(動作未確認)
