+++ app/build.gradle
@@ -36,7 +36,10 @@ dependencies {
     testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:25.1.1'
     compile 'com.fasterxml.jackson.core:jackson-databind:2.7.0'
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
+    compile 'com.github.kittinunf.fuel:fuel:1.12.1'
+    compile 'com.github.kittinunf.fuel:fuel-moshi:1.12.1'
+    compile 'com.squareup.moshi:moshi:1.5.0'
+    compile 'com.squareup.moshi:moshi-kotlin:1.5.0'
 }
 repositories {
     mavenCentral()
