+++ app/src/main/java/pakutoma/miowidget/exception/UndefinedPlanException.kt
@@ -1,0 +1,6 @@
+package pakutoma.miowidget.exception
+
+/**
+ * Created by PAKUTOMA on 2018/02/20.
+ */
+class UndefinedPlanException(str: String) : Exception(str)
