+++ app/src/main/java/pakutoma/miowidget/service/ChangeCoupon.kt
@@ -4,6 +4,7 @@ import android.app.IntentService
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
+import pakutoma.miowidget.R
 
 import pakutoma.miowidget.utility.CouponAPI
 import pakutoma.miowidget.exception.NotFoundValidTokenException
@@ -16,12 +17,20 @@ import pakutoma.miowidget.exception.NotFoundValidTokenException
 class ChangeCoupon : IntentService("ChangeCoupon") {
 
     override fun onHandleIntent(intent: Intent?) {
+        val preferences = applicationContext.getSharedPreferences("iijmio_token",Context.MODE_PRIVATE)
         try {
-            val coupon = CouponAPI(applicationContext)
-            val changedSwitch = coupon.changeCouponStatus()
+            val developerID = resources.getText(R.string.developerID).toString()
+            val accessToken = preferences.getString("X-IIJmio-Authorization","")
+            if(accessToken == "") {
+                throw NotFoundValidTokenException("Not found token in preference.")
+            }
+            val coupon = CouponAPI(developerID,accessToken)
+            val couponInfo = coupon.fetchCouponInfo()
+            val isOn = !couponInfo.lineInfoList.first().couponUse
+            val serviceCodeList = couponInfo.lineInfoList.map{it.serviceCode}
+            val changedSwitch = coupon.changeCouponUse(isOn,serviceCodeList)
             sendCallback(true, true, changedSwitch)
         } catch (e: NotFoundValidTokenException) {
-            val preferences = getSharedPreferences("iijmio_token", Context.MODE_PRIVATE)
             val editor = preferences.edit()
             editor.putString("X-IIJmio-Authorization", "")
             editor.putBoolean("has_token", false)
