+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">mioWidget</string>
     <string name="appwidget_text">未取得</string>
+    <string name="developerID">IilCI1xrAgqKrXV9Zt4</string>
 </resources>
